example: add more documentation
Currently there is not much documentation for how to use this app. This is
a problem if someone doesn't completely understand how to use the app.
This adds some very basic documentation for using this in `models.py`,
`views.py`, and a template. Helps to not have another issue like #51.